requirements.txt
selenium>=4.0.0
webdriver-manager>=3.0.0
pillow>=9.0.0